PHPStan: Add extension for typing Registry::call()
`Registry::call()` is widely used in SimplePie to allow extensibility. Unfortunately, the pattern hides the actual method from PHPStan so it cannot determine the return value type, and there is no annotation we can add to make it work. , This results in an inadequate analysis.

Thankfully, most cases allow us to determine the method statically so we can create a PHPStan extension doing just that.

The extension might be useful for SimplePie consumers so let’s expose it in `composer.json`.

We may want to add more type checking extensions in the future.